requirements.txt
python-telegram-bot==13.5
requests==2.25.1